Add reset.js - query to DB Move reset email to reset.js
